Create README.md
This program identifies the best fit value for a diffusivity for a particular set of boundary conditions, given initial condition, known time-span, and current or observed condition.  Two methods are used to accomplish this.  First, I use an analytical expression of the diffusion equation in Fourier space.  If the initial condition is known and has a Fourier transform (as most simple configurations do), then the Fourier transform at any later time has an analytical expression.  The algorithm chooses an initial guess value for D, then uses Gauss-Newton descent to identify the value of D that leads to the best fit.  This script has validity for any medium that evolves by linear diffusion like heat transfer, molecular diffusion, hillslope evolution, and flow through a porous medium AND for which the above conditions are known.